+++ src/app/components/Card/ArtistCard.module.css
